% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssc.R}
\alias{ssc.R}
\title{Ezekiel (1929) small sample correction for r}
\usage{
ssc.R(R, N, q = 1)
}
\description{
Ezekiel (1929) small sample correction for r
}
\author{
Thom Baguley (thom.baguley@ntu.ac.uk)
}
